+++ functions/package.json
@@ -14,7 +14,7 @@
   "dependencies": {
     "express": "^4.17.1",
     "firebase": "^6.2.3-1",
-    "firebase-admin": "^8.10.0",
+    "firebase-admin": "8.10.0",
     "firebase-functions": "^3.6.1"
   },
   "devDependencies": {
