+++ functions/util/config.js
@@ -1,0 +1,12 @@
+module.exports = {
+  apiKey: "AIzaSyD8f_UjWaMviBk5aUXEv2z7QWlTOVoR7HE",
+  authDomain: "thoughtshare-1660b.firebaseapp.com",
+  databaseURL: "https://thoughtshare-1660b.firebaseio.com",
+  projectId: "thoughtshare-1660b",
+  storageBucket: "thoughtshare-1660b.appspot.com",
+  messagingSenderId: "750082543278",
+  appId: "1:750082543278:web:a7986c833ca574e764273f",
+  measurementId: "G-15944FPSBE",
+};
+//   const firebase = require("firebase"); //all these require data is been installed as a package in the functions cd - npm insatall --save blah
+//   firebase.initializeApp(firebaseConfig);
