+++ functions/util/validators.js
@@ -1,0 +1,59 @@
+//helper methods to validate if fields are empty or incorrect
+const isEmail = (email) => {
+  const regEx = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
+  if (email.match(regEx)) {
+    return true;
+  } else {
+    return false;
+  }
+};
+const isEmpty = (string) => {
+  if (string.trim() === "") {
+    return true;
+  } else {
+    return false;
+  }
+};
+
+exports.validateSignupData = (data) => {
+  let errors = {};
+  if (isEmpty(data.email)) {
+    errors.email = "Must not be empty";
+  } else if (!isEmail(data.email)) {
+    errors.email = "Must be a valid email address";
+  }
+  if (isEmpty(data.password)) {
+    errors.password = "Must not be empty";
+  }
+  if (data.password !== data.confirmPassword) {
+    errors.confirmPassword = "Passwords must match";
+  }
+  if (isEmpty(data.handle)) {
+    errors.handle = "Must not be empty";
+  }
+  // validate user - checks if user handle is unique or no - this is done by checking the "users collection of db. we check if it has that handle or new user has come from below code"
+  //   if (Object.keys(errors).length > 0) {
+  //     return res.status(400).json(errors);
+  //   }
+  return {
+    errors,
+    valid: Object.keys(errors).length === 0 ? true : false,
+  };
+};
+
+exports.validateLoginData = (data) => {
+  let errors = {};
+  if (isEmpty(data.email)) {
+    errors.email = "Must not be empty";
+  }
+  //   else if (!isEmail(newUser.email)) {
+  //     errors.email = "Must be a valid email address";
+  //   }
+  if (isEmpty(data.password)) {
+    errors.password = "Must not be empty";
+  }
+  return {
+    errors,
+    valid: Object.keys(errors).length === 0 ? true : false,
+  };
+};
